Create a HTML Report About R Scripts

Create a HTML Report about each R script below a root directory. The report will contain an overview plot showing the number or rows for each script. In addition, one plot is generated per script, showing for each function defined in the script, the number of expressions contained in the function.

report_about_r_scripts(root, scripts = dir(root, "\\.R$", ignore.case =
  TRUE, recursive = TRUE), show = TRUE)

Arguments

root
path to directory from which to start looking for R scripts

scripts
optional. Paths to R scripts, relative to the path given in root

show
if TRUE the created HTML file is opened in your default browser

Value

path to the created HTML file, invisibly