How Often Are the Functions of a Package Used?

How Often Are the Functions of a Package Used?

get_package_function_usage(tree, package, simple = FALSE,
  by_script = FALSE)

Arguments

tree
parse tree as returned by parse_scripts

package
name of the package (must be installed)

simple
if TRUE, a simple approach using a simple regular expression is used. This approach is fast but not correct as it e.g. counts function calls that are commented out or even string expressions that just look like function calls. Leaving this argument to its default, FALSE, will return only real function calls by evaluating the full structure of parse tree.

by_script
if TRUE the functions are counted and returned by script, otherwise they are counted over all scripts

Value

data frame with columns name (name of the function), prefixed (number of function calls prefixed with <package>:: or <package>:::), non_prefixed (number of function calls that are not prefixed with the package name) and total (total number of function calls)

Examples

# Read all scripts that are provided in the kwb.fakin package
tree <- kwb.code::parse_scripts(root = system.file(package = "kwb.fakin"))

#> Reading /usr/local/lib/R/site-library/kwb.fakin/doc/analyse_scripts.R ... ok. (0.00s) 
#> Reading /usr/local/lib/R/site-library/kwb.fakin/doc/track_folder_changes.R ... ok. (0.00s) 
#> Reading /usr/local/lib/R/site-library/kwb.fakin/doc/understanding_encoding.R ... ok. (0.00s) 
#> Reading /usr/local/lib/R/site-library/kwb.fakin/doc/what_is_there.R ... ok. (0.00s) 
#> Reading /usr/local/lib/R/site-library/kwb.fakin/doc/working_with_paths.R ... ok. (0.00s) 
#> Reading /usr/local/lib/R/site-library/kwb.fakin/extdata/example_sizeable_sankey.R ... ok. (0.00s) 
#> Reading /usr/local/lib/R/site-library/kwb.fakin/extdata/scripts/create_treemaps.R ... ok. (0.00s) 
#> Reading /usr/local/lib/R/site-library/kwb.fakin/extdata/scripts/extract_properties.R ... ok. (0.00s) 
#> Reading /usr/local/lib/R/site-library/kwb.fakin/extdata/scripts/plot_folder_depths.R ... ok. (0.00s) 
#> Reading /usr/local/lib/R/site-library/kwb.fakin/extdata/scripts/read_different_path_files.R ... ok. (0.00s) 
#> Reading /usr/local/lib/R/site-library/kwb.fakin/extdata/scripts/test_networkD3.R ... ok. (0.00s) 
#> Reading /usr/local/lib/R/site-library/kwb.fakin/extdata/scripts/test_package.R ... ok. (0.00s) 
#> Reading /usr/local/lib/R/site-library/kwb.fakin/extdata/scripts/test_plot_files_in_depth.R ... ok. (0.00s) 
#> Reading /usr/local/lib/R/site-library/kwb.fakin/extdata/scripts/test_reporting.R ... ok. (0.00s) 
#> Reading /usr/local/lib/R/site-library/kwb.fakin/extdata/scripts/test_treemap.R ... ok. (0.00s) 
#> Reading /usr/local/lib/R/site-library/kwb.fakin/extdata/scripts/use_pathlist_class.R ... ok. (0.00s) 
#> Reading /usr/local/lib/R/site-library/kwb.fakin/extdata/scripts/write_example_file_info.R ... ok. (0.00s) 
#> Reading /usr/local/lib/R/site-library/kwb.fakin/extdata/scripts/write_file_property_availability.R ... ok. (0.00s) 
#> Reading /usr/local/lib/R/site-library/kwb.fakin/extdata/testcalls_1.R ... ok. (0.00s) 
#> Reading /usr/local/lib/R/site-library/kwb.fakin/extdata/testcalls_2.R ... ok. (0.00s) 
#> Reading /usr/local/lib/R/site-library/kwb.fakin/extdata/testcalls_3.R ... ok. (0.00s) 

# Check which functions from kwb.utils are used and how often
get_package_function_usage(tree, package = "kwb.utils")

#>      package                  name count explicit implicit
#> 1  kwb.utils              safePath     7        1        6
#> 2  kwb.utils               desktop     6        2        4
#> 3  kwb.utils     noFactorDataFrame     4        4        0
#> 4  kwb.utils              catLines     3        1        2
#> 5  kwb.utils   asNoFactorDataFrame     2        2        0
#> 6  kwb.utils            loadObject     2        2        0
#> 7  kwb.utils  assignPackageObjects     1        1        0
#> 8  kwb.utils             collapsed     1        1        0
#> 9  kwb.utils           excludeNULL     1        1        0
#> 10 kwb.utils      extractSubstring     1        1        0
#> 11 kwb.utils         fileExtension     1        1        0
#> 12 kwb.utils          getAttribute     1        1        0
#> 13 kwb.utils hsOpenWindowsExplorer     1        1        0
#> 14 kwb.utils           isNaOrEmpty     1        1        0
#> 15 kwb.utils            makeUnique     1        1        0
#> 16 kwb.utils              rbindAll     1        1        0
#> 17 kwb.utils       removeExtension     1        1        0
#> 18 kwb.utils         resetRowNames     1        1        0
#> 19 kwb.utils       toKeysAndValues     1        1        0
#> 20 kwb.utils                  user     1        1        0
#> 21 kwb.utils             writeText     1        1        0

# Hm, this does not seem to be the whole truth...