How Often Are the Functions of a Package Used?

How Often Are the Functions of a Package Used?

get_package_function_usage(tree, package, simple = FALSE,
  by_script = FALSE)

Arguments

tree
parse tree as returned by parse_scripts

package
name of the package (must be installed)

simple
if TRUE, a simple approach using a simple regular expression is used. This approach is fast but not correct as it e.g. counts function calls that are commented out or even string expressions that just look like function calls. Leaving this argument to its default, FALSE, will return only real function calls by evaluating the full structure of parse tree.

by_script
if TRUE the functions are counted and returned by script, otherwise they are counted over all scripts

Value

data frame with columns name (name of the function), prefixed (number of function calls prefixed with <package>:: or <package>:::), non_prefixed (number of function calls that are not prefixed with the package name) and total (total number of function calls)

Examples

# Read all scripts that are provided in the kwb.fakin package
tree <- kwb.code::parse_scripts(root = system.file(package = "kwb.fakin"))

#> Reading /home/travis/build/KWB-R/kwb.fakin/inst/extdata/scripts/create_treemaps.R ... ok. (0.00s) 
#> Reading /home/travis/build/KWB-R/kwb.fakin/inst/extdata/scripts/extract_properties.R ... ok. (0.00s) 
#> Reading /home/travis/build/KWB-R/kwb.fakin/inst/extdata/scripts/plot_folder_depths.R ... ok. (0.00s) 
#> Reading /home/travis/build/KWB-R/kwb.fakin/inst/extdata/scripts/summarise_file_info.R ... ok. (0.00s) 
#> Reading /home/travis/build/KWB-R/kwb.fakin/inst/extdata/scripts/test_get_file_info.R ... ok. (0.00s) 
#> Reading /home/travis/build/KWB-R/kwb.fakin/inst/extdata/scripts/test_networkD3.R ... ok. (0.00s) 
#> Reading /home/travis/build/KWB-R/kwb.fakin/inst/extdata/scripts/test_package.R ... ok. (0.00s) 
#> Reading /home/travis/build/KWB-R/kwb.fakin/inst/extdata/scripts/test_reporting.R ... ok. (0.00s) 
#> Reading /home/travis/build/KWB-R/kwb.fakin/inst/extdata/scripts/test_tree_packages.R ... ok. (0.00s) 
#> Reading /home/travis/build/KWB-R/kwb.fakin/inst/extdata/scripts/test_treemap.R ... ok. (0.00s) 
#> Reading /home/travis/build/KWB-R/kwb.fakin/inst/extdata/scripts/write_file_property_availability.R ... ok. (0.00s) 
#> Reading /home/travis/build/KWB-R/kwb.fakin/inst/extdata/scripts/write_paths_to_folder_tree.R ... ok. (0.00s) 
#> Reading /home/travis/build/KWB-R/kwb.fakin/inst/extdata/testcalls_1.R ... ok. (0.00s) 
#> Reading /home/travis/build/KWB-R/kwb.fakin/inst/extdata/testcalls_2.R ... ok. (0.00s) 
#> Reading /home/travis/build/KWB-R/kwb.fakin/inst/extdata/testcalls_3.R ... ok. (0.00s) 

# Check which functions from kwb.utils are used and how often
get_package_function_usage(tree, package = "kwb.utils")

#>      package                  name count explicit implicit
#> 1  kwb.utils              safePath    11        4        7
#> 2  kwb.utils               desktop     8        3        5
#> 3  kwb.utils              catLines     5        2        3
#> 4  kwb.utils     noFactorDataFrame     5        5        0
#> 5  kwb.utils   asNoFactorDataFrame     3        3        0
#> 6  kwb.utils         fileExtension     3        3        0
#> 7  kwb.utils          pasteColumns     3        3        0
#> 8  kwb.utils hsOpenWindowsExplorer     2        2        0
#> 9  kwb.utils            rStylePath     2        2        0
#> 10 kwb.utils  assignPackageObjects     1        1        0
#> 11 kwb.utils             catAndRun     1        1        0
#> 12 kwb.utils             collapsed     1        1        0
#> 13 kwb.utils                encode     1        1        0
#> 14 kwb.utils      extractSubstring     1        1        0
#> 15 kwb.utils          getAttribute     1        1        0
#> 16 kwb.utils         insertColumns     1        1        0
#> 17 kwb.utils           isNaOrEmpty     1        1        0
#> 18 kwb.utils            makeUnique     1        1        0
#> 19 kwb.utils              rbindAll     1        1        0
#> 20 kwb.utils       removeExtension     1        1        0
#> 21 kwb.utils         resetRowNames     1        1        0
#> 22 kwb.utils       toKeysAndValues     1        1        0

# Hm, this does not seem to be the whole truth...