Read Files Containing File Path Information from Folder

Read Files Containing File Path Information from Folder

read_path_information(file_info_dir, pattern = "^path-info", sep = ",",
  ...)

Arguments

file_info_dir
path to folder in which to look for files matching pattern

pattern
pattern to match against the file names to be read. Default: "^path-info"

sep
column separator passed to read_file_info

…
further arguments passed to read_file_info

Examples

# Set root directory (here: package installation directory of kwb.fakin)
root_dir <- system.file(package = "kwb.fakin")

# Set output directory
output_dir <- tempdir()

# Write all paths below root_dir into a "path-info"-file
kwb.fakin::get_and_save_file_info(root_dir, output_dir)

#> Getting file information on files below /home/travis/build/KWB-R/kwb.fakin/inst ... ok. (0.00s) 
#> Elapsed: 0.003 
#> Writing to '/tmp/Rtmpf5ohR3/path-info_2019-03-21_2143_inst.csv' with data.table::fwrite() ... ok. (0.00s) 
#> Elapsed: 0.001 

# Read the "path-info"-files that are (now) found in output_dir
path_info <- kwb.fakin:::read_path_information(output_dir, sep = ";")

#> Reading '/tmp/Rtmpf5ohR3/path-info_2019-03-21_2143_inst.csv' with data.table::fread() ... ok. (0.00s)